Document incompatibility with catbox-redis
Since catbox-redis does not allow storing of the integer 0 at the moment (hapijs/catbox-redis#59), the plugin does not work with catbox-redis following the first request. It fails when trying to retrieve the request from the catbox cache.

One way to workaround this, would be to start counting at 1 and not at 0. But this seems rather hacky to me.